# go: reduce GC using a soft memory limit