@jinyus 2023-10-08
Hey, thanks for the PR... but total time including I/O has been removed from the table so optimizing it is no longer necessary. It will now show how the program scales when running with 25k and 50k posts. 

Targeting specific hardware is against the [rules](https://github.com/jinyus/related_post_gen#run-benchmark) btw.