@felixge 2023-10-08
> Hey, thanks for the PR... but total time including I/O has been removed from the table so optimizing it is no longer necessary. It will now show how the program scales when running with 25k and 50k posts.

Ok, I guess you were moving the goal post as I was working on this 😅 .

> Targeting specific hardware is against the [rules](https://github.com/jinyus/related_post_gen#run-benchmark) btw.

Can you elaborate on what you mean by this? You are running your benchmarks on specific hardware, and compilers/JITs/runtimes/etc. might take advantage of this at both compile time as well as runtime. So what exactly is allowed and not allowed here?

FWIW, I don't think the `GOAMD64` flag is going to make a difference for the inner hot loop in Go. But generally speaking, I'd say taking advantage of the available hardware seems fair game?