@jinyus 2023-10-08
> Ok, I guess you were moving the goal post as I was working on this 😅 .

Yeah sorry about that,

The hardware targeting rule was originally put in place to prevent simd optimizations on x86 cpus. It just goes against the spirit of the competition.

Re the good commits:

Did you see a change when disabling arenas? I will merge the tests