@felixge 2023-10-09
> The hardware targeting rule was originally put in place to prevent simd optimizations on x86 cpus. It's just goes against the spirit of the competition.

Ok, thanks for clarifying.

> Re the good commits:
> 
> Did you see a change when disabling arenas? I will merge the tests

arenes were removed 4 days ago in fbe6f2e7380a1a0426dc6957f7578b39b121f2b1. My commit is just cleaning up something that the other commit had forgotten.